requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pyupbit
python-dotenv
pydantic
requests
sqlalchemy>=2.0.0
